refactor(DTO): DTO 패키지 재분류
