+++ server/src/main/java/kr/re/dslab/threatmodeling/controller/AttackController.java
@@ -3,7 +3,7 @@ package kr.re.dslab.threatmodeling.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.re.dslab.threatmodeling.service.AttackService;
-import kr.re.dslab.threatmodeling.type.dto.AttackRelatedResponseDto;
+import kr.re.dslab.threatmodeling.type.dto.response.AttackRelatedResponseDto;
 import kr.re.dslab.threatmodeling.type.dto.ResponseModel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
