+++ server/src/main/java/kr/re/dslab/threatmodeling/controller/MitigationController.java
@@ -3,7 +3,7 @@ package kr.re.dslab.threatmodeling.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.re.dslab.threatmodeling.service.MitigationService;
-import kr.re.dslab.threatmodeling.type.dto.MitigationResponseDto;
+import kr.re.dslab.threatmodeling.type.dto.response.MitigationResponseDto;
 import kr.re.dslab.threatmodeling.type.dto.ResponseModel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
