+++ server/src/main/java/kr/re/dslab/threatmodeling/repository/DefendRepositoryCustom.java
@@ -1,6 +1,6 @@
 package kr.re.dslab.threatmodeling.repository;
 
-import kr.re.dslab.threatmodeling.type.dto.DefendResponseDto;
+import kr.re.dslab.threatmodeling.type.dto.response.DefendResponseDto;
 
 import java.util.List;
