+++ server/src/main/java/kr/re/dslab/threatmodeling/repository/MitigationRepositoryCustom.java
@@ -1,6 +1,6 @@
 package kr.re.dslab.threatmodeling.repository;
 
-import kr.re.dslab.threatmodeling.type.dto.MitigationResponseDto;
+import kr.re.dslab.threatmodeling.type.dto.response.MitigationResponseDto;
 
 import java.util.List;
