+++ server/src/main/java/kr/re/dslab/threatmodeling/service/ControlService.java
@@ -1,7 +1,7 @@
 package kr.re.dslab.threatmodeling.service;
 
 import kr.re.dslab.threatmodeling.repository.ControlRepository;
-import kr.re.dslab.threatmodeling.type.dto.ControlResponseDto;
+import kr.re.dslab.threatmodeling.type.dto.response.ControlResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
