+++ server/src/main/java/kr/re/dslab/threatmodeling/service/DataInitializationService.java
@@ -11,10 +11,10 @@ import kr.re.dslab.threatmodeling.repository.CveRepository;
 import kr.re.dslab.threatmodeling.repository.DefendRepository;
 import kr.re.dslab.threatmodeling.repository.MitigationAttackMappingRepository;
 import kr.re.dslab.threatmodeling.repository.MitigationRepository;
-import kr.re.dslab.threatmodeling.type.dto.ControlAttackAttackDto;
-import kr.re.dslab.threatmodeling.type.dto.ControlAttackDto;
-import kr.re.dslab.threatmodeling.type.dto.MitigationAttackDto;
-import kr.re.dslab.threatmodeling.type.dto.MitigationDefendDto;
+import kr.re.dslab.threatmodeling.type.dto.json.ControlAttackAttackDto;
+import kr.re.dslab.threatmodeling.type.dto.json.ControlAttackDto;
+import kr.re.dslab.threatmodeling.type.dto.json.MitigationAttackDto;
+import kr.re.dslab.threatmodeling.type.dto.json.MitigationDefendDto;
 import kr.re.dslab.threatmodeling.type.entity.Attack;
 import kr.re.dslab.threatmodeling.type.entity.Control;
 import kr.re.dslab.threatmodeling.type.entity.ControlAttackMapping;
