+++ server/src/main/java/kr/re/dslab/threatmodeling/service/MitigationService.java
@@ -3,8 +3,8 @@ package kr.re.dslab.threatmodeling.service;
 import kr.re.dslab.threatmodeling.exception.NotFoundException;
 import kr.re.dslab.threatmodeling.repository.DefendRepository;
 import kr.re.dslab.threatmodeling.repository.MitigationRepository;
-import kr.re.dslab.threatmodeling.type.dto.DefendResponseDto;
-import kr.re.dslab.threatmodeling.type.dto.MitigationResponseDto;
+import kr.re.dslab.threatmodeling.type.dto.response.DefendResponseDto;
+import kr.re.dslab.threatmodeling.type.dto.response.MitigationResponseDto;
 import kr.re.dslab.threatmodeling.type.entity.Mitigation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
