+++ server/src/main/java/kr/re/dslab/threatmodeling/type/dto/json/ControlAttackDto.java
@@ -1,4 +1,4 @@
-package kr.re.dslab.threatmodeling.type.dto;
+package kr.re.dslab.threatmodeling.type.dto.json;
 
 import kr.re.dslab.threatmodeling.type.entity.Control;
 import lombok.Data;
