+++ server/src/main/java/kr/re/dslab/threatmodeling/type/dto/response/DefendResponseDto.java
@@ -1,4 +1,4 @@
-package kr.re.dslab.threatmodeling.type.dto;
+package kr.re.dslab.threatmodeling.type.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
