+++ server/src/main/java/kr/re/dslab/threatmodeling/type/entity/Mitigation.java
@@ -3,8 +3,8 @@ package kr.re.dslab.threatmodeling.type.entity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
-import kr.re.dslab.threatmodeling.type.dto.MitigationAttackDto;
-import kr.re.dslab.threatmodeling.type.dto.MitigationDefendDto;
+import kr.re.dslab.threatmodeling.type.dto.json.MitigationAttackDto;
+import kr.re.dslab.threatmodeling.type.dto.json.MitigationDefendDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
